quest_1/requirements.txt
fastapi[all]~=0.109.0
python-dotenv~=1.0.0
sqlalchemy~=2.0.25
psycopg2
alembic
pydantic~=2.5.3